backend/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
reportlab>=4.0.4
django-allauth>=0.54.0
dj-rest-auth>=4.0.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
django-environ>=0.10.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
Pillow>=9.5.0
whitenoise>=6.5.0
python-dotenv>=1.0.0
requests>=2.31.0
stripe>=6.0.0
pdfkit>=1.0.0
django-filter>=23.2
cryptography>=41.0.0
jinja2>=3.1.2
